backend/requirements.txt
flask
flask-cors
pyrebase4
gunicorn
